first draft of river visualization site!
This first draft should include basic branding to match the current riverml.xyz site
and a simple example with river and plotting centroids. I will need to turn on gh-pages for
the automated docs workflow to work, and I will also be adding circle-ci soon to have previews
on pull requests.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>